requirements.txt
streamlit>=1.22.0
opencv-python>=4.7.0
numpy>=1.22.0
matplotlib>=3.5.0
pillow>=9.0.0
ipywidgets>=7.6.0
scikit-image>=0.18.0 